requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1
    # via django
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
click==7.1.2
    # via pip-tools
cryptography==3.4.2
    # via pyjwt
defusedxml==0.6.0
    # via python3-openid
django-allauth==0.44.0
    # via -r requirements.in
django-ordered-model==3.4.1
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-allauth
django_crispy_forms==1.11.0
    # via -r requirements.in
idna==2.10
    # via requests
oauthlib==3.1.0
    # via requests-oauthlib
pip-tools==5.5.0
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.0.1
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via django
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
sqlparse==0.4.1
    # via django
urllib3==1.26.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
